Started with implementing how attacks hit
Added empty game objects to both hands, both calfs and both feet of the player prefabs. Added a gameobject array with hit origins in player. This array contains references to the empties mentioned earlier. changed HitboxDriver with DoHit(). Takes a string and splits it with ',' as delimiter. The string notates which origins the attack will use. Called from animation events.